refactored : net_connection_interface
